+++ sdk/storage/azure-storage-blobs/inc/azure/storage/blobs/blob_client.hpp
@@ -273,6 +273,15 @@ namespace Azure { namespace Storage { namespace Blobs {
         const DeleteBlobOptions& options = DeleteBlobOptions()) const;
 
     /**
+     * @brief Marks the specified blob or snapshot for deletion if it exists.
+     *
+     * @param options Optional parameters to execute this function.
+     * @return A DeleteBlobResult on successfully deleting. Null if the blob doesn't exist.
+     */
+    Azure::Core::Response<Models::DeleteBlobResult> DeleteIfExists(
+        const DeleteBlobOptions& options = DeleteBlobOptions()) const;
+
+    /**
      * @brief Restores the contents and metadata of a soft deleted blob and any associated
      * soft deleted snapshots.
      *